Add go.mod to pixel-examples
As-is stops working in Go 1.15.
This was fixed with "go mod init" and "go mod tidy".
There doesn't seem to be particular value in per-example go.mod files, so I've made one for the entire project.

I wasn't able to get community/amidakuji/glossary/util.go to build; it seems the example may be unfinished.
Otherwise, the examples I tried seemed to function as intended.

Add explicit instructions for Windows systems.
It took me hours of trying various configurations to find one which works on a fresh machine, this should help new users.

Add /.idea and *.exe to .gitignore to preserve my sanity in testing this.